rag-service/requirements.txt
# rag-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.25.0
pydantic>=2.5.2
python-dotenv>=1.0.0
numpy>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
groq>=0.4.0
watchdog>=3.0.0
aiofiles>=23.2.1
python-multipart>=0.0.6
pyyaml>=6.0
httpx>=0.25.0